+++ src/modules/controls/freeze/tests/end2end/Freeze.e2e.js
@@ -15,7 +15,7 @@ const webdriver = require("selenium-webdriver"),
  * @param {module:selenium-webdriver.Capabilities} param.capability sets the capability when requesting a new session - overwrites all previously set capabilities
  * @returns {void}
  */
-function FreezeTests ({builder, url, resolution, browsername, capability}) {
+async function FreezeTests ({builder, url, resolution, browsername, capability}) {
     const testIsApplicable = !isMobile(resolution) && // function not available mobile
         (isCustom(url) || isMaster(url)); // freeze only active in these